+++ CHANGELOG.md
@@ -29,6 +29,7 @@ Types of changes:
 # 2.14.1 - 2020-08-17
 ### Changed
 - add possibility to hide 'Modify URL' button
+- add confirm modal on click delete image
 
 # 2.14.0 - 2020-08-17
 ### Changed